algo: python3-compatible NiceSort
NiceSort relies on the fact that, in Python 2,

     None < Int < String

However, cmp() under Python3 refuses to compare None w/ Int w/ Strings,
so we need to rework NiceSort's logic to only use keys of the same type.

We convert all keys to strings, with None being mapped to the empty
string and integer values being zero-padded to 10 digits. This yields
the same results as the implicit Python 2 ordering, as long as integer
parts remain below 10 digits.