Do not use ConfigObjects as set members
ConfigObjects are mutable and thus unhashable. Python 3 requires that
objects implementing __eq__() also hash to the same value, but it is
impossible to have a stable hash value for mutable objects like this.

Instead, we just face reality that ConfigObjects are mutable, and they
can't be used as keys or set members for the same reasons dicts cannot
be used as well.